refactor(scaffolding): split ScaffoldingService.cs into three partial files by scaffold type
ScaffoldingService.cs (2776 lines) is split into focused partial-class files:
- ScaffoldingService.TypePreview.cs: type scaffolding + interface resolution helpers
- ScaffoldingService.TestPreview.cs: single-test scaffolding + sibling-pattern inference
- ScaffoldingService.TestBatchAndFirstTestPreview.cs: batch-test and first-test-file scaffolding

Pure code organization — no behavior changes. All 24 ScaffoldingIntegration tests pass unchanged.

Closes backlog row: scaffolding-service-split-by-scaffold-type
