+++ src/RoslynMcp.Roslyn/Services/ScaffoldingService.cs
@@ -19,7 +19,7 @@ internal static class MinimalSymbolDisplayExtensions
         symbol.ToDisplayString(SymbolDisplayFormat.MinimallyQualifiedFormat);
 }
 
-public sealed class ScaffoldingService : IScaffoldingService
+public sealed partial class ScaffoldingService : IScaffoldingService
 {
     private readonly IWorkspaceManager _workspace;
     private readonly IFileOperationService _fileOperationService;
@@ -33,30 +33,6 @@ public sealed class ScaffoldingService : IScaffoldingService
         _workspace = workspace;
         _fileOperationService = fileOperationService;
         _previewStore = previewStore;
-    }
-
-    public async Task<RefactoringPreviewDto> PreviewScaffoldTypeAsync(string workspaceId, ScaffoldTypeDto request, CancellationToken ct)
-    {
-        IdentifierValidation.ThrowIfInvalidIdentifier(request.TypeName, "type name");
-        var project = ResolveProject(workspaceId, request.ProjectName);
-        var projectDirectory = Path.GetDirectoryName(project.FilePath)
-            ?? throw new InvalidOperationException($"Project directory could not be resolved for '{project.FilePath}'.");
-        var typeNamespace = string.IsNullOrWhiteSpace(request.Namespace) ? project.Name : request.Namespace!;
-        var folderSegments = ResolveFolderSegmentsForNamespace(typeNamespace, project.Name);
-        var filePath = Path.Combine([projectDirectory, .. folderSegments, $"{request.TypeName}.cs"]);
-
-        var interfaceResolution = await ResolveInterfaceMembersAsync(workspaceId, request, ct).ConfigureAwait(false);
-        var content = BuildTypeContent(typeNamespace, request, interfaceResolution);
-
-        var preview = await _fileOperationService
-            .PreviewCreateFileAsync(workspaceId, new CreateFileDto(project.Name, filePath, content), ct)
-            .ConfigureAwait(false);
-
-        if (interfaceResolution.Warnings.Count > 0)
-        {
-            return preview with { Warnings = interfaceResolution.Warnings };
-        }
-        return preview;
     }
 
     /// <summary>
@@ -80,189 +56,6 @@ public sealed class ScaffoldingService : IScaffoldingService
             : typeNamespace;
 
         return workingNamespace.Split('.', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
-    }
-
-    /// <summary>
-    /// Item 8: batch scaffold-test. Runs the single-type resolver once per target but reuses a
-    /// single workspace solution snapshot across all targets, aggregating the document adds into
-    /// one <see cref="RefactoringPreviewDto"/>/preview token. Callers redeem the token with
-    /// <c>apply_composite_preview</c> (or the regular apply path) to commit the whole batch
-    /// atomically.
-    /// </summary>
-    public async Task<RefactoringPreviewDto> PreviewScaffoldTestBatchAsync(
-        string workspaceId, ScaffoldTestBatchDto request, CancellationToken ct)
-    {
-        ValidateBatchScaffoldRequest(request);
-        var context = CreateBatchScaffoldContext(workspaceId, request);
-        var cachedCompilations = await LoadBatchCompilationsAsync(context.Solution, context.TestProject, ct).ConfigureAwait(false);
-        var state = new BatchScaffoldState(context.Solution);
-
-        foreach (var target in request.Targets)
-        {
-            ProcessBatchScaffoldTarget(target, request, context, cachedCompilations, state, ct);
-        }
-
-        return await CreateBatchScaffoldPreviewAsync(workspaceId, context.Project, state, ct).ConfigureAwait(false);
-    }
-
-    private static void ValidateBatchScaffoldRequest(ScaffoldTestBatchDto request)
-    {
-        if (request.Targets is null || request.Targets.Count == 0)
-        {
-            throw new InvalidOperationException("scaffold_test_batch_preview requires at least one target.");
-        }
-    }
-
-    private BatchScaffoldContext CreateBatchScaffoldContext(string workspaceId, ScaffoldTestBatchDto request)
-    {
-        var project = ResolveProject(workspaceId, request.TestProjectName);
-        ValidateIsTestProject(project);
-        var projectDirectory = Path.GetDirectoryName(project.FilePath)
-            ?? throw new InvalidOperationException($"Project directory could not be resolved for '{project.FilePath}'.");
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var testProject = solution.Projects.FirstOrDefault(p =>
-            string.Equals(p.Name, request.TestProjectName, StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(p.FilePath, request.TestProjectName, StringComparison.OrdinalIgnoreCase))
-            ?? throw new InvalidOperationException($"Test project not loaded: {request.TestProjectName}");
-
-        return new BatchScaffoldContext(
-            Project: project,
-            TestProject: testProject,
-            Solution: solution,
-            ProjectDirectory: projectDirectory,
-            TestNamespace: project.Name,
-            Framework: ResolveTestFramework(request.TestFramework, project.FilePath),
-            NSubstituteAvailable: IsNSubstituteAvailable(testProject));
-    }
-
-    private static async Task<List<Compilation>> LoadBatchCompilationsAsync(
-        Solution solution,
-        Project testProject,
-        CancellationToken ct)
-    {
-        // Cache source-project compilations once to avoid N× GetCompilationAsync across targets
-        // (the primary perf win over iterating PreviewScaffoldTestAsync).
-        var projectsToSearch = new List<Project> { testProject };
-        foreach (var projectRef in testProject.ProjectReferences)
-        {
-            var referenced = solution.GetProject(projectRef.ProjectId);
-            if (referenced is not null)
-            {
-                projectsToSearch.Add(referenced);
-            }
-        }
-
-        var cachedCompilations = new List<Compilation>();
-        foreach (var project in projectsToSearch)
-        {
-            var compilation = await project.GetCompilationAsync(ct).ConfigureAwait(false);
-            if (compilation is not null)
-            {
-                cachedCompilations.Add(compilation);
-            }
-        }
-
-        return cachedCompilations;
-    }
-
-    private static void ProcessBatchScaffoldTarget(
-        ScaffoldTestBatchTargetDto target,
-        ScaffoldTestBatchDto request,
-        BatchScaffoldContext context,
-        IReadOnlyList<Compilation> cachedCompilations,
-        BatchScaffoldState state,
-        CancellationToken ct)
-    {
-        ct.ThrowIfCancellationRequested();
-        if (string.IsNullOrWhiteSpace(target.TargetTypeName))
-        {
-            state.Warnings.Add("Skipped empty target type name.");
-            return;
-        }
-
-        // See PreviewScaffoldTestAsync — accept dotted FQN input, resolve via simple name,
-        // and let the matched symbol supply the authoritative class identifier.
-        var lookupName = StripToSimpleTypeName(target.TargetTypeName);
-        var typeInfo = ResolveTargetTypeAndMethodFromCache(
-            cachedCompilations,
-            lookupName,
-            target.TargetMethodName,
-            context.NSubstituteAvailable);
-        if (typeInfo.MatchedType is null)
-        {
-            state.Warnings.Add($"Target type '{target.TargetTypeName}' not found in referenced projects — skipped.");
-            return;
-        }
-
-        var simpleTypeName = typeInfo.MatchedType.Name;
-        var testFilePath = Path.Combine(context.ProjectDirectory, $"{simpleTypeName}GeneratedTests.cs");
-        if (SymbolResolver.FindDocument(state.Accumulator, testFilePath) is not null || File.Exists(testFilePath))
-        {
-            state.Warnings.Add($"Skipped '{target.TargetTypeName}': target file already exists at '{testFilePath}'.");
-            return;
-        }
-
-        if (typeInfo.Warnings is not null)
-        {
-            state.Warnings.AddRange(typeInfo.Warnings);
-        }
-
-        var dto = new ScaffoldTestDto(
-            request.TestProjectName,
-            target.TargetTypeName,
-            target.TargetMethodName,
-            request.TestFramework);
-
-        // Batch scaffolding intentionally does NOT apply sibling-pattern inference — a batch
-        // run typically targets a homogenous set of production types and callers want the
-        // generic scaffold. Sibling inference is available via per-target scaffold_test_preview.
-        var content = BuildTestContent(
-            context.TestNamespace,
-            dto,
-            simpleTypeName,
-            typeInfo.TargetNamespace,
-            typeInfo.ConstructorArgs,
-            context.Framework,
-            typeInfo.TargetMethod,
-            typeInfo.MatchedType,
-            siblingPattern: null,
-            isTargetInaccessible: typeInfo.IsTargetInaccessible);
-
-        var testProject = state.Accumulator.GetProject(context.TestProject.Id)
-            ?? throw new InvalidOperationException("Test project disappeared from working solution snapshot.");
-        var newDocument = testProject.AddDocument(
-            Path.GetFileName(testFilePath),
-            Microsoft.CodeAnalysis.Text.SourceText.From(content),
-            folders: [],
-            filePath: testFilePath);
-
-        state.Accumulator = newDocument.Project.Solution;
-        state.CreatedFiles.Add(testFilePath);
-    }
-
-    private async Task<RefactoringPreviewDto> CreateBatchScaffoldPreviewAsync(
-        string workspaceId,
-        ProjectStatusDto project,
-        BatchScaffoldState state,
-        CancellationToken ct)
-    {
-        if (state.CreatedFiles.Count == 0)
-        {
-            throw new InvalidOperationException(
-                "scaffold_test_batch_preview produced no file creations. See Warnings for per-target reasons.");
-        }
-
-        var changes = await Helpers.SolutionDiffHelper
-            .ComputeChangesAsync(state.OriginalSolution, state.Accumulator, ct)
-            .ConfigureAwait(false);
-        var description = $"Scaffold {state.CreatedFiles.Count} test file(s) in project '{project.Name}'";
-        var token = _previewStore.Store(workspaceId, state.Accumulator, _workspace.GetCurrentVersion(workspaceId), description);
-
-        return new RefactoringPreviewDto(
-            token,
-            description,
-            changes,
-            state.Warnings.Count > 0 ? state.Warnings : null);
     }
 
     private sealed record BatchScaffoldContext(
@@ -301,663 +94,6 @@ public sealed class ScaffoldingService : IScaffoldingService
     {
         public static ResolvedTargetTypeInfo NotFound { get; } = new(string.Empty, string.Empty, null, null, null);
     }
-
-    /// <summary>
-    /// Non-async variant of <see cref="ResolveTargetTypeAndMethodAsync"/> reading from a cached
-    /// compilation list. Used by batch scaffold to avoid re-walking <c>GetCompilationAsync</c>
-    /// per target.
-    /// </summary>
-    private static ResolvedTargetTypeInfo
-        ResolveTargetTypeAndMethodFromCache(
-            IReadOnlyList<Compilation> compilations,
-            string targetTypeName,
-            string? targetMethodName,
-            bool nsubstituteAvailable = false)
-    {
-        INamedTypeSymbol? matchedType = null;
-        foreach (var compilation in compilations)
-        {
-            var candidates = GetMatchingTargetTypeCandidates(compilation, targetTypeName, CancellationToken.None).ToList();
-            if (candidates.Count == 1) { matchedType = candidates[0]; break; }
-            if (candidates.Count > 1)
-            {
-                return CreateAmbiguousTargetTypeResult(targetTypeName);
-            }
-        }
-
-        return CreateResolvedTargetTypeInfo(matchedType, targetMethodName, warnOnPrivateMethod: false, nsubstituteAvailable);
-    }
-
-    public async Task<RefactoringPreviewDto> PreviewScaffoldTestAsync(
-        string workspaceId,
-        ScaffoldTestDto request,
-        CancellationToken ct,
-        ITestNameSuggestionProvider? testNameSuggestionProvider = null)
-    {
-        var project = ResolveProject(workspaceId, request.TestProjectName);
-        ValidateIsTestProject(project);
-        var projectDirectory = Path.GetDirectoryName(project.FilePath)
-            ?? throw new InvalidOperationException($"Project directory could not be resolved for '{project.FilePath}'.");
-        var testNamespace = project.Name;
-
-        var framework = ResolveTestFramework(request.TestFramework, project.FilePath);
-
-        // Accept a dotted FQN as input (callers who hit the ambiguity error get pointed at
-        // "the fully qualified type name", then re-invoke with `Namespace.Type`). The resolver
-        // only ever looks up the simple name, so strip to that for lookup — and treat the
-        // matched symbol's Name as authoritative once we have it, so the downstream class
-        // identifier is always a single identifier (dotted identifiers are a CS syntax error).
-        var lookupName = StripToSimpleTypeName(request.TargetTypeName);
-        var typeInfo = await ResolveTargetTypeAndMethodAsync(
-            workspaceId, request.TestProjectName, lookupName, request.TargetMethodName, ct).ConfigureAwait(false);
-
-        var simpleTypeName = typeInfo.MatchedType?.Name ?? lookupName;
-        var testFilePath = Path.Combine(projectDirectory, $"{simpleTypeName}GeneratedTests.cs");
-
-        // Sibling-pattern inference (scaffold-test-sibling-pattern-inference). When an explicit
-        // referenceTestFile is supplied we use that as the pattern source; otherwise we
-        // auto-detect the most-recently-modified `*Tests.cs` in the project directory. Empty
-        // string opts out of inference.
-        // Per scaffold-test-preview-sibling-inference-overbroad: we pass the test project's
-        // compilation so usings can be trimmed to those actually referenced by the captured
-        // surface (base list + ctor params). Without semantic resolution the scaffold pulls in
-        // every using from the sibling fixture (typically 10+ unused imports).
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var testRoslynProject = solution.Projects.FirstOrDefault(p =>
-            string.Equals(p.Name, project.Name, StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(p.FilePath, project.FilePath, StringComparison.OrdinalIgnoreCase));
-        var testProjectCompilation = testRoslynProject is null
-            ? null
-            : await testRoslynProject.GetCompilationAsync(ct).ConfigureAwait(false);
-        var siblingInference = InferSiblingTestPattern(request.ReferenceTestFile, projectDirectory, testFilePath, testProjectCompilation);
-        var siblingWarnings = siblingInference.Warnings;
-        var sampledTestName = await SuggestSampledTestNameAsync(
-            request,
-            simpleTypeName,
-            typeInfo.TargetNamespace,
-            typeInfo.TargetMethod,
-            projectDirectory,
-            testFilePath,
-            testNameSuggestionProvider,
-            ct).ConfigureAwait(false);
-
-        var content = BuildTestContent(
-            testNamespace, request, simpleTypeName, typeInfo.TargetNamespace, typeInfo.ConstructorArgs, framework,
-            typeInfo.TargetMethod, typeInfo.MatchedType, siblingInference.Pattern, sampledTestName.MethodName,
-            isTargetInaccessible: typeInfo.IsTargetInaccessible);
-        var preview = await _fileOperationService.PreviewCreateFileAsync(workspaceId, new CreateFileDto(project.Name, testFilePath, content), ct).ConfigureAwait(false);
-
-        var combinedWarnings = CombineWarnings(typeInfo.Warnings, siblingWarnings, sampledTestName.Warning);
-        return combinedWarnings.Count == 0 ? preview : preview with { Warnings = combinedWarnings };
-    }
-
-    private static async Task<TestNameSuggestionResult> SuggestSampledTestNameAsync(
-        ScaffoldTestDto request,
-        string simpleTypeName,
-        string targetNamespace,
-        IMethodSymbol? targetMethod,
-        string projectDirectory,
-        string testFilePath,
-        ITestNameSuggestionProvider? provider,
-        CancellationToken ct)
-    {
-        if (!request.UseSampling || string.IsNullOrWhiteSpace(request.TargetMethodName))
-        {
-            return new TestNameSuggestionResult(null);
-        }
-
-        if (provider is null)
-        {
-            return new TestNameSuggestionResult(
-                null,
-                "useSampling was true but no sampling provider was available; emitted the deterministic placeholder test name.");
-        }
-
-        try
-        {
-            var context = new ScaffoldTestNameSuggestionContext(
-                simpleTypeName,
-                request.TargetMethodName,
-                FormatMethodSignature(targetMethod),
-                string.IsNullOrWhiteSpace(targetNamespace) ? null : targetNamespace,
-                CollectSiblingTestMethodNames(projectDirectory, testFilePath, maxNames: 6));
-            var result = await provider.SuggestTestNameAsync(context, ct).ConfigureAwait(false);
-            var normalized = NormalizeSuggestedTestMethodName(result.MethodName);
-            if (normalized is not null)
-            {
-                return result with { MethodName = normalized };
-            }
-
-            return string.IsNullOrWhiteSpace(result.MethodName)
-                ? new TestNameSuggestionResult(null, result.Warning)
-                : new TestNameSuggestionResult(
-                    null,
-                    AppendWarning(result.Warning, $"Sampled test method name '{result.MethodName}' was not a valid C# identifier; emitted the deterministic placeholder test name."));
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            return new TestNameSuggestionResult(
-                null,
-                $"Sampling test-name suggestion failed ({ex.GetType().Name}: {ex.Message}); emitted the deterministic placeholder test name.");
-        }
-    }
-
-    /// <summary>
-    /// Previews scaffolding the FIRST <c>&lt;Service&gt;Tests.cs</c> file for a service that
-    /// has no existing fixture in the destination test project. Resolves the service via
-    /// fully-qualified <see cref="ScaffoldFirstTestFileDto.ServiceMetadataName"/> (so two services
-    /// with the same simple name in different namespaces are unambiguous), infers the
-    /// destination test project when not supplied, captures the boilerplate shape from up to
-    /// three most-recently-modified <c>*Tests.cs</c> sibling fixtures, and emits a fixture with
-    /// one smoke-test per public method on the service.
-    /// </summary>
-    public async Task<RefactoringPreviewDto> PreviewScaffoldFirstTestFileAsync(
-        string workspaceId, ScaffoldFirstTestFileDto request, CancellationToken ct)
-    {
-        if (string.IsNullOrWhiteSpace(request.ServiceMetadataName))
-            throw new InvalidOperationException("scaffold_first_test_file_preview requires a non-empty serviceMetadataName.");
-
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var (serviceSymbol, sourceProject) = await ResolveServiceByMetadataNameAsync(solution, request.ServiceMetadataName, ct).ConfigureAwait(false);
-        if (serviceSymbol is null || sourceProject is null)
-        {
-            throw new InvalidOperationException(
-                $"Service '{request.ServiceMetadataName}' was not found by metadata name in any loaded project. " +
-                "Pass the fully-qualified type name (Namespace.TypeName).");
-        }
-
-        var simpleTypeName = serviceSymbol.Name;
-        IdentifierValidation.ThrowIfInvalidIdentifier(simpleTypeName, "service type name");
-
-        var testProject = ResolveDestinationTestProject(workspaceId, request.TestProjectName, sourceProject);
-        ValidateIsTestProject(testProject);
-        var projectDirectory = Path.GetDirectoryName(testProject.FilePath)
-            ?? throw new InvalidOperationException($"Project directory could not be resolved for '{testProject.FilePath}'.");
-
-        var testFilePath = Path.Combine(projectDirectory, $"{simpleTypeName}Tests.cs");
-        if (File.Exists(testFilePath))
-        {
-            throw new InvalidOperationException(
-                $"Destination file '{testFilePath}' already exists. " +
-                "scaffold_first_test_file_preview is for brand-new fixtures only — use scaffold_test_preview to add tests to an existing fixture.");
-        }
-
-        var framework = ResolveTestFramework(request.TestFramework, testProject.FilePath);
-        var testRoslynProject = solution.Projects.FirstOrDefault(p =>
-            string.Equals(p.Name, testProject.Name, StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(p.FilePath, testProject.FilePath, StringComparison.OrdinalIgnoreCase));
-        // Per scaffold-test-preview-sibling-inference-overbroad: pass the test project's
-        // compilation to trim usings semantically.
-        var testCompilation = testRoslynProject is null
-            ? null
-            : await testRoslynProject.GetCompilationAsync(ct).ConfigureAwait(false);
-        var siblingInference = InferSiblingPatternFromRecent(projectDirectory, testFilePath, maxSiblings: 3, testCompilation);
-
-        var serviceNamespace = serviceSymbol.ContainingNamespace.IsGlobalNamespace
-            ? string.Empty
-            : serviceSymbol.ContainingNamespace.ToDisplayString();
-        var nsubstituteAvailable = IsNSubstituteAvailable(testRoslynProject);
-        var constructorArgs = BuildConstructorArgs(serviceSymbol, nsubstituteAvailable);
-        var publicMethods = CollectPublicTestableMethods(serviceSymbol);
-
-        var content = BuildFirstTestFileContent(
-            testProject.Name,
-            serviceNamespace,
-            simpleTypeName,
-            constructorArgs,
-            publicMethods,
-            framework,
-            siblingInference.Pattern,
-            serviceSymbol);
-
-        var warnings = new List<string>();
-        warnings.AddRange(siblingInference.Warnings);
-        if (publicMethods.Count == 0)
-        {
-            warnings.Add(
-                $"Service '{simpleTypeName}' has no public/internal instance methods to scaffold smoke tests for — emitted a single placeholder test.");
-        }
-
-        var preview = await _fileOperationService
-            .PreviewCreateFileAsync(workspaceId, new CreateFileDto(testProject.Name, testFilePath, content), ct)
-            .ConfigureAwait(false);
-
-        return warnings.Count == 0 ? preview : preview with { Warnings = warnings };
-    }
-
-    private static async Task<(INamedTypeSymbol? Symbol, Project? Project)> ResolveServiceByMetadataNameAsync(
-        Solution solution, string serviceMetadataName, CancellationToken ct)
-    {
-        // Parse the metadata name into "Namespace.Simple" pieces so we can fall back to
-        // GetSymbolsWithName(simpleName) when GetTypeByMetadataName fails. The fast-path
-        // (GetTypeByMetadataName) is the common case; the fallbacks handle transient
-        // compilation-state gaps where the type-by-metadata-name lookup returns null for
-        // a type that is still reachable via the slower symbol-enumeration path — observed
-        // under fresh-load + immediate-query timing in MSBuild-backed test workspaces.
-        var simpleName = serviceMetadataName.Contains('.', StringComparison.Ordinal)
-            ? serviceMetadataName[(serviceMetadataName.LastIndexOf('.') + 1)..]
-            : serviceMetadataName;
-
-        // Walk every project, but accept ONLY when the symbol is declared in source within
-        // that project (not merely visible via a project/metadata reference). Using
-        // ContainingAssembly identity means we land on the owning project — critical for
-        // ResolveDestinationTestProject which walks ProjectReferences from the OWNER.
-        foreach (var project in solution.Projects)
-        {
-            var compilation = await project.GetCompilationAsync(ct).ConfigureAwait(false);
-            if (compilation is null) continue;
-
-            // Fast path.
-            var byMetadataName = compilation.GetTypeByMetadataName(serviceMetadataName);
-            if (byMetadataName is { TypeKind: TypeKind.Class } &&
-                SymbolEqualityComparer.Default.Equals(byMetadataName.ContainingAssembly, compilation.Assembly))
-            {
-                return (byMetadataName, project);
-            }
-
-            // Fallback 1: enumerate source-declared types with the simple name and pick the
-            // one whose full display string matches. GetSymbolsWithName walks declared syntax
-            // names, which can return results even in edge cases where the metadata-name
-            // lookup has not yet published the type to the global cache.
-            if (string.IsNullOrEmpty(simpleName)) continue;
-            var candidates = compilation.GetSymbolsWithName(simpleName, SymbolFilter.Type, ct)
-                .OfType<INamedTypeSymbol>()
-                .Where(t => t.TypeKind == TypeKind.Class &&
-                            SymbolEqualityComparer.Default.Equals(t.ContainingAssembly, compilation.Assembly) &&
-                            string.Equals(t.ToDisplayString(), serviceMetadataName, StringComparison.Ordinal))
-                .ToList();
-            if (candidates.Count == 1)
-            {
-                return (candidates[0], project);
-            }
-
-            // Fallback 2: walk the project's syntax trees directly for a matching class
-            // declaration. Only used when both symbol-index paths miss — typically because
-            // the compilation's symbol table has not yet caught up with a freshly-loaded
-            // MSBuildWorkspace. Uses the SemanticModel on the tree that contains a
-            // matching class declaration, which forces that tree's symbols to materialize.
-            var syntaxFound = await FindByDeclaredClassSyntaxAsync(project, compilation, simpleName, serviceMetadataName, ct).ConfigureAwait(false);
-            if (syntaxFound is not null)
-            {
-                return (syntaxFound, project);
-            }
-        }
-        return (null, null);
-    }
-
-    private static async Task<INamedTypeSymbol?> FindByDeclaredClassSyntaxAsync(
-        Project project, Compilation compilation, string simpleName, string fullMetadataName, CancellationToken ct)
-    {
-        foreach (var document in project.Documents)
-        {
-            ct.ThrowIfCancellationRequested();
-            var tree = await document.GetSyntaxTreeAsync(ct).ConfigureAwait(false);
-            if (tree is null) continue;
-            var root = await tree.GetRootAsync(ct).ConfigureAwait(false);
-            var classDecls = root.DescendantNodes()
-                .OfType<Microsoft.CodeAnalysis.CSharp.Syntax.ClassDeclarationSyntax>()
-                .Where(c => string.Equals(c.Identifier.ValueText, simpleName, StringComparison.Ordinal));
-            foreach (var decl in classDecls)
-            {
-                var model = compilation.GetSemanticModel(tree);
-                var symbol = model.GetDeclaredSymbol(decl, ct);
-                if (symbol is INamedTypeSymbol named &&
-                    string.Equals(named.ToDisplayString(), fullMetadataName, StringComparison.Ordinal))
-                {
-                    return named;
-                }
-            }
-        }
-        return null;
-    }
-
-    /// <summary>
-    /// Picks the destination test project. When the caller supplies a name/path, we resolve
-    /// directly. When omitted, we look for a project that (a) references the source production
-    /// project and (b) has a project name ending in <c>.Tests</c> — the canonical convention
-    /// used across this repo and most .NET solutions. Throws when no candidate is found.
-    /// </summary>
-    private ProjectStatusDto ResolveDestinationTestProject(string workspaceId, string? requestedName, Project sourceProject)
-    {
-        if (!string.IsNullOrWhiteSpace(requestedName))
-            return ResolveProject(workspaceId, requestedName);
-
-        var status = _workspace.GetStatus(workspaceId);
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var sourceId = sourceProject.Id;
-
-        var candidates = solution.Projects
-            .Where(p => p.ProjectReferences.Any(r => r.ProjectId == sourceId))
-            .Where(p => p.Name.EndsWith(".Tests", StringComparison.Ordinal))
-            .ToList();
-
-        if (candidates.Count == 0)
-        {
-            throw new InvalidOperationException(
-                $"Could not infer a destination test project for service in project '{sourceProject.Name}'. " +
-                "No project ending in '.Tests' references this project. Pass testProjectName explicitly.");
-        }
-
-        if (candidates.Count > 1)
-        {
-            var names = string.Join(", ", candidates.Select(c => c.Name));
-            throw new InvalidOperationException(
-                $"Multiple test projects reference '{sourceProject.Name}': {names}. Pass testProjectName explicitly.");
-        }
-
-        var picked = candidates[0];
-        return status.Projects.FirstOrDefault(p => string.Equals(p.Name, picked.Name, StringComparison.OrdinalIgnoreCase))
-            ?? throw new InvalidOperationException($"Inferred test project '{picked.Name}' is not in workspace status — cannot resolve file path.");
-    }
-
-    /// <summary>
-    /// Returns public + internal instance methods declared on the service (excluding inherited
-    /// <see cref="object"/> members, property accessors, constructors, and operators) — the set
-    /// we emit smoke tests for in a first-test-file scaffold.
-    /// </summary>
-    private static IReadOnlyList<IMethodSymbol> CollectPublicTestableMethods(INamedTypeSymbol service)
-    {
-        return service.GetMembers()
-            .OfType<IMethodSymbol>()
-            .Where(m => m.MethodKind == MethodKind.Ordinary)
-            .Where(m => !m.IsStatic)
-            .Where(m => m.DeclaredAccessibility is Accessibility.Public or Accessibility.Internal)
-            .Where(m => m.AssociatedSymbol is null) // skip property/event accessors
-            .ToList();
-    }
-
-    /// <summary>
-    /// Inspects up to <paramref name="maxSiblings"/> most-recently-modified <c>*Tests.cs</c>
-    /// fixtures in <paramref name="projectDirectory"/> and returns the pattern captured from
-    /// the most recent one (used as the "primary" boilerplate template). Falls back to
-    /// repo-convention defaults (returns <see cref="SiblingInferenceResult.None"/>) when no
-    /// sibling fixtures exist — the per-framework emitter then writes the framework's standard
-    /// boilerplate.
-    /// <para>
-    /// When <paramref name="compilation"/> is non-null, captured <c>using</c> directives are
-    /// trimmed via semantic resolution to only those required by the captured surface. See
-    /// <c>scaffold-test-preview-sibling-inference-overbroad</c>.
-    /// </para>
-    /// </summary>
-    private static SiblingInferenceResult InferSiblingPatternFromRecent(
-        string projectDirectory, string destinationFilePath, int maxSiblings, Compilation? compilation = null)
-    {
-        if (!Directory.Exists(projectDirectory))
-            return SiblingInferenceResult.None;
-
-        var destinationNormalized = Path.GetFullPath(destinationFilePath);
-        var siblings = Directory.EnumerateFiles(projectDirectory, "*Tests.cs", SearchOption.AllDirectories)
-            .Where(p =>
-            {
-                var normalized = Path.GetFullPath(p);
-                if (string.Equals(normalized, destinationNormalized, StringComparison.OrdinalIgnoreCase))
-                    return false;
-                var rel = Path.GetRelativePath(projectDirectory, normalized);
-                return !rel.Split(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
-                    .Any(seg => string.Equals(seg, "obj", StringComparison.OrdinalIgnoreCase)
-                             || string.Equals(seg, "bin", StringComparison.OrdinalIgnoreCase));
-            })
-            .Select(p => new FileInfo(p))
-            .OrderByDescending(fi => fi.LastWriteTimeUtc)
-            .Take(maxSiblings)
-            .ToList();
-
-        if (siblings.Count == 0)
-            return SiblingInferenceResult.None;
-
-        var warnings = new List<string>();
-        // Use the most-recently-modified sibling as the primary pattern source. The maxSiblings
-        // window above is a forward-compatible hook — today only the freshest sibling drives the
-        // scaffolded shape; future revisions can union attribute lists across the window.
-        try
-        {
-            var primary = siblings[0];
-            var sourceText = File.ReadAllText(primary.FullName);
-            var pattern = ExtractPatternFromSource(sourceText, primary.Name, compilation, primary.FullName);
-            return new SiblingInferenceResult(pattern, warnings);
-        }
-        catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-        {
-            warnings.Add(
-                $"Could not read sibling fixture '{siblings[0].Name}' ({ex.GetType().Name}: {ex.Message}) — scaffolded without pattern inference.");
-            return new SiblingInferenceResult(null, warnings);
-        }
-    }
-
-    /// <summary>
-    /// Builds a brand-new <c>&lt;Service&gt;Tests.cs</c> fixture: emits the framework header,
-    /// a class-level <c>ClassInitialize</c> hook (MSTest only — xUnit/NUnit get equivalent
-    /// constructor/<c>OneTimeSetUp</c> patterns), a single <c>subject</c> field of the service
-    /// type wired up via the inferred constructor args, and one smoke-test method per public
-    /// instance method. Sibling fragments (extra usings, class attributes, base class) are
-    /// layered onto the default template.
-    /// </summary>
-    private static string BuildFirstTestFileContent(
-        string testNamespace,
-        string serviceNamespace,
-        string serviceTypeName,
-        string constructorArgs,
-        IReadOnlyList<IMethodSymbol> publicMethods,
-        string framework,
-        SiblingTestPattern? siblingPattern,
-        INamedTypeSymbol? serviceSymbol = null)
-    {
-        var usingDirective = string.IsNullOrWhiteSpace(serviceNamespace)
-            ? string.Empty
-            : $"using {serviceNamespace};\n";
-
-        // Collect namespaces from constructor parameter types that differ from the service's own
-        // namespace — these are not captured by the single usingDirective above and would cause
-        // CS0246 when the generated file is compiled as-is (scaffold-test-preview-missing-usings).
-        var ctorParamUsings = BuildCtorParamUsings(serviceSymbol, serviceNamespace, testNamespace);
-
-        var ctorCall = string.IsNullOrWhiteSpace(constructorArgs)
-            ? $"new {serviceTypeName}()"
-            : $"new {serviceTypeName}({constructorArgs})";
-
-        return framework switch
-        {
-            "xunit" => BuildFirstTestFileXunit(testNamespace, usingDirective, ctorParamUsings, serviceTypeName, ctorCall, publicMethods, siblingPattern),
-            "nunit" => BuildFirstTestFileNUnit(testNamespace, usingDirective, ctorParamUsings, serviceTypeName, ctorCall, publicMethods, siblingPattern),
-            _ => BuildFirstTestFileMSTest(testNamespace, usingDirective, ctorParamUsings, serviceTypeName, ctorCall, publicMethods, siblingPattern),
-        };
-    }
-
-    private static string BuildFirstTestFileMSTest(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string serviceTypeName,
-        string ctorCall,
-        IReadOnlyList<IMethodSymbol> publicMethods,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, _ctorBlock) =
-            BuildSiblingFragments(siblingPattern, serviceTypeName, testNamespace, usingDirective);
-
-        // First-test-file skips the sibling ctor-block: ClassInitialize takes the role of
-        // shared setup, and we don't want to inherit a fixture-injection ctor from a sibling
-        // that happens to use IClassFixture<T>. The base-class clause is preserved so users
-        // who base their fixtures on SharedWorkspaceTestBase keep that linkage.
-        var fixtureName = serviceTypeName + "Tests";
-        var sb = new System.Text.StringBuilder();
-        sb.Append("using Microsoft.VisualStudio.TestTools.UnitTesting;\n");
-        sb.Append(usingDirective);
-        sb.Append(ctorParamUsings);
-        sb.Append(extraUsings);
-        sb.Append('\n');
-        sb.Append("namespace ").Append(testNamespace).Append(";\n\n");
-        sb.Append(extraAttributes);
-        sb.Append("[TestClass]\n");
-        sb.Append("public sealed class ").Append(fixtureName).Append(baseClause).Append('\n');
-        sb.Append("{\n");
-        sb.Append("    private static ").Append(serviceTypeName).Append("? _subject;\n\n");
-        sb.Append("    [ClassInitialize]\n");
-        sb.Append("    public static void ClassInit(TestContext _)\n");
-        sb.Append("    {\n");
-        sb.Append("        _subject = ").Append(ctorCall).Append(";\n");
-        sb.Append("    }\n\n");
-
-        if (publicMethods.Count == 0)
-        {
-            sb.Append("    [TestMethod]\n");
-            sb.Append("    public void Subject_Is_Constructible()\n");
-            sb.Append("    {\n");
-            sb.Append("        Assert.IsNotNull(_subject);\n");
-            sb.Append("    }\n");
-        }
-        else
-        {
-            for (var i = 0; i < publicMethods.Count; i++)
-            {
-                var method = publicMethods[i];
-                sb.Append("    [TestMethod]\n");
-                sb.Append("    public void ").Append(method.Name).Append("_Smoke_Needs_Real_Test()\n");
-                sb.Append("    {\n");
-                sb.Append("        Assert.IsNotNull(_subject);\n");
-                sb.Append("        // TODO: invoke _subject!.").Append(method.Name).Append("(...) and assert.\n");
-                sb.Append("    }\n");
-                if (i < publicMethods.Count - 1) sb.Append('\n');
-            }
-        }
-
-        sb.Append("}\n");
-        return sb.ToString();
-    }
-
-    private static string BuildFirstTestFileXunit(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string serviceTypeName,
-        string ctorCall,
-        IReadOnlyList<IMethodSymbol> publicMethods,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, _ctorBlock) =
-            BuildSiblingFragments(siblingPattern, serviceTypeName, testNamespace, usingDirective);
-
-        var fixtureName = serviceTypeName + "Tests";
-        var sb = new System.Text.StringBuilder();
-        sb.Append("using Xunit;\n");
-        sb.Append(usingDirective);
-        sb.Append(ctorParamUsings);
-        sb.Append(extraUsings);
-        sb.Append('\n');
-        sb.Append("namespace ").Append(testNamespace).Append(";\n\n");
-        sb.Append(extraAttributes);
-        sb.Append("public sealed class ").Append(fixtureName).Append(baseClause).Append('\n');
-        sb.Append("{\n");
-        sb.Append("    private readonly ").Append(serviceTypeName).Append(" _subject;\n\n");
-        sb.Append("    public ").Append(fixtureName).Append("()\n");
-        sb.Append("    {\n");
-        sb.Append("        _subject = ").Append(ctorCall).Append(";\n");
-        sb.Append("    }\n\n");
-
-        if (publicMethods.Count == 0)
-        {
-            sb.Append("    [Fact]\n");
-            sb.Append("    public void Subject_Is_Constructible()\n");
-            sb.Append("    {\n");
-            sb.Append("        Assert.NotNull(_subject);\n");
-            sb.Append("    }\n");
-        }
-        else
-        {
-            for (var i = 0; i < publicMethods.Count; i++)
-            {
-                var method = publicMethods[i];
-                sb.Append("    [Fact]\n");
-                sb.Append("    public void ").Append(method.Name).Append("_Smoke_Needs_Real_Test()\n");
-                sb.Append("    {\n");
-                sb.Append("        Assert.NotNull(_subject);\n");
-                sb.Append("        // TODO: invoke _subject.").Append(method.Name).Append("(...) and assert.\n");
-                sb.Append("    }\n");
-                if (i < publicMethods.Count - 1) sb.Append('\n');
-            }
-        }
-
-        sb.Append("}\n");
-        return sb.ToString();
-    }
-
-    private static string BuildFirstTestFileNUnit(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string serviceTypeName,
-        string ctorCall,
-        IReadOnlyList<IMethodSymbol> publicMethods,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, _ctorBlock) =
-            BuildSiblingFragments(siblingPattern, serviceTypeName, testNamespace, usingDirective);
-
-        var fixtureName = serviceTypeName + "Tests";
-        var sb = new System.Text.StringBuilder();
-        sb.Append("using NUnit.Framework;\n");
-        sb.Append(usingDirective);
-        sb.Append(ctorParamUsings);
-        sb.Append(extraUsings);
-        sb.Append('\n');
-        sb.Append("namespace ").Append(testNamespace).Append(";\n\n");
-        sb.Append(extraAttributes);
-        sb.Append("[TestFixture]\n");
-        sb.Append("public sealed class ").Append(fixtureName).Append(baseClause).Append('\n');
-        sb.Append("{\n");
-        sb.Append("    private ").Append(serviceTypeName).Append("? _subject;\n\n");
-        sb.Append("    [OneTimeSetUp]\n");
-        sb.Append("    public void OneTimeSetUp()\n");
-        sb.Append("    {\n");
-        sb.Append("        _subject = ").Append(ctorCall).Append(";\n");
-        sb.Append("    }\n\n");
-
-        if (publicMethods.Count == 0)
-        {
-            sb.Append("    [Test]\n");
-            sb.Append("    public void Subject_Is_Constructible()\n");
-            sb.Append("    {\n");
-            sb.Append("        Assert.That(_subject, Is.Not.Null);\n");
-            sb.Append("    }\n");
-        }
-        else
-        {
-            for (var i = 0; i < publicMethods.Count; i++)
-            {
-                var method = publicMethods[i];
-                sb.Append("    [Test]\n");
-                sb.Append("    public void ").Append(method.Name).Append("_Smoke_Needs_Real_Test()\n");
-                sb.Append("    {\n");
-                sb.Append("        Assert.That(_subject, Is.Not.Null);\n");
-                sb.Append("        // TODO: invoke _subject!.").Append(method.Name).Append("(...) and assert.\n");
-                sb.Append("    }\n");
-                if (i < publicMethods.Count - 1) sb.Append('\n');
-            }
-        }
-
-        sb.Append("}\n");
-        return sb.ToString();
-    }
-
-    private static IReadOnlyList<string> CombineWarnings(List<string>? a, IReadOnlyList<string>? b, string? c = null)
-    {
-        if ((a is null || a.Count == 0) && (b is null || b.Count == 0) && string.IsNullOrWhiteSpace(c))
-            return Array.Empty<string>();
-        var combined = new List<string>();
-        if (a is not null) combined.AddRange(a);
-        if (b is not null) combined.AddRange(b);
-        if (!string.IsNullOrWhiteSpace(c)) combined.Add(c);
-        return combined;
-    }
-
-    private static string AppendWarning(string? existingWarning, string warning)
-        => string.IsNullOrWhiteSpace(existingWarning)
-            ? warning
-            : $"{existingWarning} {warning}";
 
     private static string ResolveTestFramework(string? requested, string? projectFilePath)
     {
@@ -1000,85 +136,6 @@ public sealed class ScaffoldingService : IScaffoldingService
         }
 
         return "mstest";
-    }
-
-    private async Task<ResolvedTargetTypeInfo>
-        ResolveTargetTypeAndMethodAsync(
-            string workspaceId, string testProjectName, string targetTypeName, string? targetMethodName, CancellationToken ct)
-    {
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var testProject = solution.Projects.FirstOrDefault(p =>
-            string.Equals(p.Name, testProjectName, StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(p.FilePath, testProjectName, StringComparison.OrdinalIgnoreCase));
-
-        if (testProject is null)
-            return ResolvedTargetTypeInfo.NotFound;
-
-        var projectsToSearch = GetProjectsToSearch(solution, testProject);
-        var matchedType = await FindTargetTypeAsync(projectsToSearch, targetTypeName, ct).ConfigureAwait(false);
-        var nsubstituteAvailable = IsNSubstituteAvailable(testProject);
-
-        // scaffold-test-internal-target-accessibility: when the target type/method is internal
-        // and the test assembly lacks InternalsVisibleTo, the previous output produced
-        // direct `new TargetType()` / `subject.Method()` calls that fail compile with CS0122.
-        // Surface this as a warning + non-applicable scaffold so callers can decide between
-        // adding InternalsVisibleTo, moving the target to public surface, or scaffolding from
-        // a project that already has access.
-        var testCompilation = await testProject.GetCompilationAsync(ct).ConfigureAwait(false);
-        var testAssembly = testCompilation?.Assembly;
-        return CreateResolvedTargetTypeInfo(
-            matchedType,
-            targetMethodName,
-            warnOnPrivateMethod: true,
-            nsubstituteAvailable,
-            testAssembly,
-            testProject.Name);
-    }
-
-    private static List<Project> GetProjectsToSearch(Solution solution, Project testProject)
-    {
-        var projectsToSearch = new List<Project> { testProject };
-        foreach (var projectRef in testProject.ProjectReferences)
-        {
-            var referencedProject = solution.GetProject(projectRef.ProjectId);
-            if (referencedProject is not null)
-            {
-                projectsToSearch.Add(referencedProject);
-            }
-        }
-
-        return projectsToSearch;
-    }
-
-    private static async Task<INamedTypeSymbol?> FindTargetTypeAsync(
-        IReadOnlyList<Project> projectsToSearch,
-        string targetTypeName,
-        CancellationToken ct)
-    {
-        foreach (var project in projectsToSearch)
-        {
-            var compilation = await project.GetCompilationAsync(ct).ConfigureAwait(false);
-            if (compilation is null)
-            {
-                continue;
-            }
-
-            var candidates = GetMatchingTargetTypeCandidates(compilation, targetTypeName, ct).ToList();
-            if (candidates.Count == 1)
-            {
-                return candidates[0];
-            }
-
-            if (candidates.Count > 1)
-            {
-                throw new InvalidOperationException(
-                    $"Ambiguous type name '{targetTypeName}' — found in multiple namespaces: " +
-                    string.Join(", ", candidates.Select(c => c.ToDisplayString())) +
-                    ". Use the fully qualified type name.");
-            }
-        }
-
-        return null;
     }
 
     private static IEnumerable<INamedTypeSymbol> GetMatchingTargetTypeCandidates(
@@ -1478,273 +535,6 @@ public sealed class ScaffoldingService : IScaffoldingService
             new(string.Empty, Array.Empty<string>(), Array.Empty<string>());
     }
 
-    /// <summary>
-    /// Item 2: when <see cref="ScaffoldTypeDto.ImplementInterface"/> is true and the scaffolded
-    /// type is a class (not an interface/record/enum), walk any interface candidates in
-    /// <c>BaseType</c> and <c>Interfaces</c>, resolve each to an <see cref="INamedTypeSymbol"/>,
-    /// and build textual stub declarations for all interface members. Falls back to
-    /// <see cref="InterfaceResolutionResult.Empty"/> (with a warning) when the interface cannot
-    /// be resolved so scaffold still succeeds.
-    /// </summary>
-    private async Task<InterfaceResolutionResult> ResolveInterfaceMembersAsync(
-        string workspaceId, ScaffoldTypeDto request, CancellationToken ct)
-    {
-        if (!request.ImplementInterface)
-            return InterfaceResolutionResult.Empty;
-
-        var normalizedKind = request.TypeKind.ToLowerInvariant();
-        if (normalizedKind is "interface" or "enum")
-            return InterfaceResolutionResult.Empty;
-
-        var candidates = CollectInterfaceCandidates(request);
-        if (candidates.Count == 0)
-            return InterfaceResolutionResult.Empty;
-
-        var solution = _workspace.GetCurrentSolution(workspaceId);
-        var stubs = new System.Text.StringBuilder();
-        var requiredUsings = new HashSet<string>(StringComparer.Ordinal);
-        var warnings = new List<string>();
-        var emittedSignatures = new HashSet<string>(StringComparer.Ordinal);
-
-        foreach (var candidate in candidates)
-        {
-            var resolved = await ResolveTypeSymbolAsync(solution, candidate, ct).ConfigureAwait(false);
-            if (resolved is null)
-            {
-                warnings.Add(
-                    $"Could not resolve '{candidate}' to a type symbol in workspace '{workspaceId}'. " +
-                    "Scaffolded class will have an empty body — add interface members manually.");
-                continue;
-            }
-
-            if (resolved.TypeKind != TypeKind.Interface)
-                continue; // concrete base class — no stubs needed.
-
-            EmitMembersForInterface(resolved, stubs, requiredUsings, emittedSignatures);
-        }
-
-        return new InterfaceResolutionResult(stubs.ToString(), requiredUsings, warnings);
-    }
-
-    /// <summary>
-    /// Collect candidate type names to resolve as interfaces: the optional <see cref="ScaffoldTypeDto.BaseType"/>
-    /// plus every non-blank entry in <see cref="ScaffoldTypeDto.Interfaces"/>. Base-type entries that turn out
-    /// to be concrete classes are filtered later in the resolution loop.
-    /// </summary>
-    private static List<string> CollectInterfaceCandidates(ScaffoldTypeDto request)
-    {
-        var candidates = new List<string>();
-        if (!string.IsNullOrWhiteSpace(request.BaseType))
-            candidates.Add(request.BaseType!);
-        if (request.Interfaces is not null)
-            candidates.AddRange(request.Interfaces.Where(n => !string.IsNullOrWhiteSpace(n)));
-        return candidates;
-    }
-
-    /// <summary>
-    /// Resolve a candidate type name across every project's compilation, preferring an exact metadata-name
-    /// match and falling back to <c>GetSymbolsWithName</c> with a display-name compare so short names
-    /// (e.g. <c>IDisposable</c>) still match without a leading namespace.
-    /// </summary>
-    private static async Task<INamedTypeSymbol?> ResolveTypeSymbolAsync(
-        Solution solution, string candidate, CancellationToken ct)
-    {
-        foreach (var project in solution.Projects)
-        {
-            var compilation = await project.GetCompilationAsync(ct).ConfigureAwait(false);
-            if (compilation is null) continue;
-            var resolved = compilation.GetTypeByMetadataName(candidate)
-                ?? compilation.GetSymbolsWithName(
-                        StripGenericArity(candidate), Microsoft.CodeAnalysis.SymbolFilter.Type, ct)
-                    .OfType<INamedTypeSymbol>()
-                    .FirstOrDefault(t => t.ToDisplayString().Equals(candidate, StringComparison.Ordinal) ||
-                                         t.Name.Equals(candidate, StringComparison.Ordinal));
-            if (resolved is not null) return resolved;
-        }
-        return null;
-    }
-
-    /// <summary>
-    /// Emit stub text for every required-to-implement member of the resolved interface and all its
-    /// inherited interfaces (<see cref="INamedTypeSymbol.AllInterfaces"/>), deduped by signature so the
-    /// same member inherited via two paths is only emitted once.
-    /// </summary>
-    private static void EmitMembersForInterface(
-        INamedTypeSymbol resolved,
-        System.Text.StringBuilder stubs,
-        HashSet<string> requiredUsings,
-        HashSet<string> emittedSignatures)
-    {
-        // AllInterfaces includes inherited interfaces (IFoo : IBar implements IBar's members too).
-        var interfacesToEmit = new List<INamedTypeSymbol> { resolved };
-        interfacesToEmit.AddRange(resolved.AllInterfaces);
-
-        foreach (var iface in interfacesToEmit)
-        {
-            foreach (var member in iface.GetMembers())
-            {
-                if (ShouldSkipInterfaceMember(member)) continue;
-
-                var signature = BuildMemberSignatureKey(member);
-                if (!emittedSignatures.Add(signature)) continue;
-
-                var stub = BuildInterfaceMemberStub(member, requiredUsings);
-                if (stub is null) continue;
-                stubs.AppendLine(stub);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Skip static interface members (DIM entry points), property/event accessors (handled via the
-    /// owning property/event itself), and members with a default implementation (C# 8 DIM — not
-    /// required of implementors).
-    /// </summary>
-    private static bool ShouldSkipInterfaceMember(ISymbol member)
-    {
-        if (member.IsStatic) return true;
-        if (member is IMethodSymbol methodSym &&
-            methodSym.AssociatedSymbol is IPropertySymbol or IEventSymbol) return true;
-        if (!member.IsAbstract && HasDefaultInterfaceImplementation(member)) return true;
-        return false;
-    }
-
-    /// <summary>Rope generic arity (<c>`1</c>) from a display name like <c>IEnumerable`1</c>.</summary>
-    private static string StripGenericArity(string name)
-    {
-        var tick = name.IndexOf('`');
-        return tick < 0 ? name : name[..tick];
-    }
-
-    /// <summary>True when an interface method/property has a concrete default body (C# 8 DIM).</summary>
-    private static bool HasDefaultInterfaceImplementation(ISymbol member)
-    {
-        foreach (var syntaxRef in member.DeclaringSyntaxReferences)
-        {
-            var node = syntaxRef.GetSyntax();
-            if (node is Microsoft.CodeAnalysis.CSharp.Syntax.MethodDeclarationSyntax m && m.Body is not null)
-                return true;
-            if (node is Microsoft.CodeAnalysis.CSharp.Syntax.PropertyDeclarationSyntax p &&
-                p.AccessorList is not null &&
-                p.AccessorList.Accessors.Any(a => a.Body is not null || a.ExpressionBody is not null))
-                return true;
-        }
-        return false;
-    }
-
-    private static string BuildMemberSignatureKey(ISymbol member)
-    {
-        if (member is IMethodSymbol method)
-        {
-            var parms = string.Join(",", method.Parameters.Select(p => p.Type.ToDisplayString()));
-            return $"M:{method.Name}({parms})<{method.TypeParameters.Length}>";
-        }
-        if (member is IPropertySymbol prop)
-        {
-            var parms = string.Join(",", prop.Parameters.Select(p => p.Type.ToDisplayString()));
-            return $"P:{prop.Name}[{parms}]";
-        }
-        return $"{member.Kind}:{member.Name}";
-    }
-
-    /// <summary>
-    /// Emit a textual method/property/event stub that throws <c>NotImplementedException</c>.
-    /// Uses <c>MinimallyQualifiedFormat</c> so callers get readable type names plus recorded
-    /// required <c>using</c> namespaces for later header assembly.
-    /// </summary>
-    private static string? BuildInterfaceMemberStub(ISymbol member, HashSet<string> requiredUsings)
-    {
-        requiredUsings.Add("System"); // NotImplementedException
-
-        if (member is IMethodSymbol method && method.MethodKind == MethodKind.Ordinary)
-        {
-            CollectNamespaces(method.ReturnType, requiredUsings);
-            foreach (var p in method.Parameters)
-                CollectNamespaces(p.Type, requiredUsings);
-
-            var typeParams = method.TypeParameters.Length == 0
-                ? string.Empty
-                : "<" + string.Join(", ", method.TypeParameters.Select(tp => tp.Name)) + ">";
-            var parameters = string.Join(", ", method.Parameters.Select(FormatParameter));
-            var constraints = BuildTypeParameterConstraints(method.TypeParameters);
-            var returnType = method.ReturnsVoid ? "void" : method.ReturnType.ToMinimalDisplay();
-            var body = method.ReturnsVoid
-                ? "throw new NotImplementedException();"
-                : "throw new NotImplementedException();";
-
-            return
-                $"    public {returnType} {method.Name}{typeParams}({parameters}){constraints}\n" +
-                "    {\n" +
-                $"        {body}\n" +
-                "    }\n";
-        }
-
-        if (member is IPropertySymbol property && !property.IsIndexer)
-        {
-            CollectNamespaces(property.Type, requiredUsings);
-            var type = property.Type.ToMinimalDisplay();
-            var accessors = new List<string>();
-            if (property.GetMethod is not null) accessors.Add("get => throw new NotImplementedException();");
-            if (property.SetMethod is not null)
-            {
-                // Use set or init based on the interface declaration.
-                var keyword = property.SetMethod.IsInitOnly ? "init" : "set";
-                accessors.Add($"{keyword} => throw new NotImplementedException();");
-            }
-            var accessorBlock = string.Join(" ", accessors);
-            return $"    public {type} {property.Name} {{ {accessorBlock} }}\n";
-        }
-
-        if (member is IEventSymbol evt)
-        {
-            CollectNamespaces(evt.Type, requiredUsings);
-            var type = evt.Type.ToMinimalDisplay();
-            return
-                $"    public event {type}? {evt.Name}\n" +
-                "    {\n" +
-                "        add => throw new NotImplementedException();\n" +
-                "        remove => throw new NotImplementedException();\n" +
-                "    }\n";
-        }
-
-        return null;
-    }
-
-    private static string FormatParameter(IParameterSymbol parameter)
-    {
-        var modifier = parameter.RefKind switch
-        {
-            RefKind.Ref => "ref ",
-            RefKind.Out => "out ",
-            RefKind.In => "in ",
-            _ => string.Empty
-        };
-        var typeText = parameter.Type.ToMinimalDisplay();
-        return parameter.IsParams
-            ? $"params {typeText} {parameter.Name}"
-            : $"{modifier}{typeText} {parameter.Name}";
-    }
-
-    private static string BuildTypeParameterConstraints(ImmutableArray<ITypeParameterSymbol> typeParameters)
-    {
-        if (typeParameters.Length == 0) return string.Empty;
-        var parts = new List<string>();
-        foreach (var tp in typeParameters)
-        {
-            var clauses = new List<string>();
-            if (tp.HasReferenceTypeConstraint) clauses.Add("class");
-            if (tp.HasValueTypeConstraint) clauses.Add("struct");
-            if (tp.HasUnmanagedTypeConstraint) clauses.Add("unmanaged");
-            if (tp.HasNotNullConstraint) clauses.Add("notnull");
-            foreach (var ct in tp.ConstraintTypes)
-                clauses.Add(ct.ToDisplayString(SymbolDisplayFormat.MinimallyQualifiedFormat));
-            if (tp.HasConstructorConstraint) clauses.Add("new()");
-            if (clauses.Count == 0) continue;
-            parts.Add($" where {tp.Name} : {string.Join(", ", clauses)}");
-        }
-        return string.Concat(parts);
-    }
-
     private static void CollectNamespaces(ITypeSymbol type, HashSet<string> requiredUsings)
     {
         if (type is null) return;
@@ -1809,55 +599,6 @@ public sealed class ScaffoldingService : IScaffoldingService
         return sb.ToString();
     }
 
-    private static string BuildTypeContent(string typeNamespace, ScaffoldTypeDto request, InterfaceResolutionResult interfaceResolution)
-    {
-        var inheritance = new List<string>();
-        if (!string.IsNullOrWhiteSpace(request.BaseType))
-        {
-            inheritance.Add(request.BaseType);
-        }
-
-        if (request.Interfaces is not null)
-        {
-            inheritance.AddRange(request.Interfaces.Where(@interface => !string.IsNullOrWhiteSpace(@interface)));
-        }
-
-        var inheritanceClause = inheritance.Count > 0 ? $" : {string.Join(", ", inheritance)}" : string.Empty;
-        var normalizedKind = request.TypeKind.ToLowerInvariant();
-        var typeKeyword = normalizedKind switch
-        {
-            "interface" => "interface",
-            "record" => "record",
-            "enum" => "enum",
-            _ => "class"
-        };
-
-        // Modern .NET convention: default scaffolded classes to `internal sealed class` so
-        // they don't expand the public API surface and aren't subclassable by accident.
-        // Records/interfaces/enums stay `public` (interface and enum cannot be sealed; records
-        // are typically intended as DTOs that get used widely).
-        var modifier = normalizedKind == "interface" || normalizedKind == "record" || normalizedKind == "enum"
-            ? "public"
-            : "internal sealed";
-
-        // Item 2: deduplicate usings against the implied namespace. Skip any using equal to the
-        // scaffolded type's own namespace — we're already in it.
-        var usingsBlock = new System.Text.StringBuilder();
-        foreach (var ns in interfaceResolution.RequiredUsings
-            .Where(n => !string.Equals(n, typeNamespace, StringComparison.Ordinal))
-            .OrderBy(n => n, StringComparer.Ordinal))
-        {
-            usingsBlock.Append("using ").Append(ns).Append(";\n");
-        }
-        if (usingsBlock.Length > 0) usingsBlock.Append('\n');
-
-        var body = string.IsNullOrEmpty(interfaceResolution.MemberStubs)
-            ? string.Empty
-            : interfaceResolution.MemberStubs;
-
-        return $"{usingsBlock}namespace {typeNamespace};\n\n{modifier} {typeKeyword} {request.TypeName}{inheritanceClause}\n{{\n{body}}}\n";
-    }
-
     /// <summary>
     /// Captured shape of a sibling test class: attributes decorating the class declaration,
     /// optional base class, and constructor-injected fixture parameters (the xUnit
@@ -1885,481 +626,6 @@ public sealed class ScaffoldingService : IScaffoldingService
     }
 
     /// <summary>
-    /// Inspects an explicit <paramref name="referenceTestFile"/> (when provided) or the
-    /// most-recently-modified <c>*Tests.cs</c> sibling in the target test project directory,
-    /// and captures its class-level attributes + base types + constructor-injected fixture
-    /// parameters into a <see cref="SiblingTestPattern"/>. Scaffold generators replicate the
-    /// captured shape so integration-test conventions carry over to the new file. A deliberate
-    /// opt-out is supported by passing an empty string for <paramref name="referenceTestFile"/>.
-    /// <para>
-    /// When <paramref name="compilation"/> is non-null, captured <c>using</c> directives are
-    /// trimmed via semantic resolution to only those required by the captured surface (base
-    /// types + constructor parameter types). See <c>scaffold-test-preview-sibling-inference-overbroad</c>.
-    /// </para>
-    /// </summary>
-    internal static SiblingInferenceResult InferSiblingTestPattern(
-        string? referenceTestFile,
-        string projectDirectory,
-        string destinationFilePath,
-        Compilation? compilation = null)
-    {
-        // Explicit opt-out: empty string means "do not infer".
-        if (referenceTestFile is not null && string.IsNullOrWhiteSpace(referenceTestFile))
-            return SiblingInferenceResult.None;
-
-        string? resolved;
-        var warnings = new List<string>();
-
-        if (!string.IsNullOrWhiteSpace(referenceTestFile))
-        {
-            if (!File.Exists(referenceTestFile))
-            {
-                warnings.Add(
-                    $"referenceTestFile '{referenceTestFile}' not found on disk — falling back to auto-detection.");
-                resolved = FindMostRecentSiblingTestFile(projectDirectory, destinationFilePath);
-            }
-            else
-            {
-                resolved = referenceTestFile;
-            }
-        }
-        else
-        {
-            resolved = FindMostRecentSiblingTestFile(projectDirectory, destinationFilePath);
-        }
-
-        if (resolved is null)
-            return new SiblingInferenceResult(null, warnings);
-
-        try
-        {
-            var sourceText = File.ReadAllText(resolved);
-            var pattern = ExtractPatternFromSource(sourceText, Path.GetFileName(resolved), compilation, resolved);
-            return new SiblingInferenceResult(pattern, warnings);
-        }
-        catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-        {
-            warnings.Add(
-                $"Could not read referenceTestFile '{resolved}' ({ex.GetType().Name}: {ex.Message}) — scaffolded without pattern inference.");
-            return new SiblingInferenceResult(null, warnings);
-        }
-    }
-
-    private static string? FindMostRecentSiblingTestFile(string projectDirectory, string destinationFilePath)
-    {
-        if (!Directory.Exists(projectDirectory))
-            return null;
-
-        // Only consider files ending in *Tests.cs (the canonical convention that scaffold
-        // itself produces). Exclude the destination so we never self-reference, and skip
-        // obj/bin subdirectories so we don't pick up generated files.
-        var destinationNormalized = Path.GetFullPath(destinationFilePath);
-        return Directory.EnumerateFiles(projectDirectory, "*Tests.cs", SearchOption.AllDirectories)
-            .Where(p =>
-            {
-                var normalized = Path.GetFullPath(p);
-                if (string.Equals(normalized, destinationNormalized, StringComparison.OrdinalIgnoreCase))
-                    return false;
-                var rel = Path.GetRelativePath(projectDirectory, normalized);
-                return !rel.Split(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
-                    .Any(seg => string.Equals(seg, "obj", StringComparison.OrdinalIgnoreCase)
-                             || string.Equals(seg, "bin", StringComparison.OrdinalIgnoreCase));
-            })
-            .Select(p => new FileInfo(p))
-            .OrderByDescending(fi => fi.LastWriteTimeUtc)
-            .Select(fi => fi.FullName)
-            .FirstOrDefault();
-    }
-
-    /// <summary>
-    /// Roslyn-parses the sibling source file, picks the first top-level class declaration
-    /// (ignoring nested types), and captures: class-level attribute lists, base type list,
-    /// constructor parameters (for <c>IClassFixture&lt;T&gt;</c> / fixture injection),
-    /// and any <c>using</c>s we'll need to carry over so the scaffolded file compiles.
-    /// Returns <c>null</c> when the file isn't parseable or has no class declaration.
-    /// <para>
-    /// Per <c>scaffold-test-preview-sibling-inference-overbroad</c>: when a non-null
-    /// <paramref name="compilation"/> is supplied, the captured <c>using</c> set is trimmed
-    /// to only those required to resolve identifiers actually referenced in the captured
-    /// surface (base list + constructor parameter types) — semantic resolution is performed
-    /// against the test project's compilation. Without this trim the scaffold pulls in every
-    /// <c>using</c> from the sibling fixture (typically 10+ unused imports when the MRU
-    /// sibling is a Playwright / Selenium fixture).
-    /// </para>
-    /// </summary>
-    private static SiblingTestPattern? ExtractPatternFromSource(
-        string sourceText,
-        string sourceFileName,
-        Compilation? compilation,
-        string? sourceFilePath)
-    {
-        var tree = CSharpSyntaxTree.ParseText(sourceText);
-        var root = tree.GetCompilationUnitRoot();
-
-        var classDecl = root.DescendantNodes()
-            .OfType<ClassDeclarationSyntax>()
-            .FirstOrDefault(c => c.Modifiers.Any(m => m.IsKind(SyntaxKind.PublicKeyword) || m.IsKind(SyntaxKind.InternalKeyword))
-                              || c.Modifiers.Count == 0);
-        if (classDecl is null)
-            return null;
-
-        var attributes = classDecl.AttributeLists
-            .Select(list => list.ToString().Trim())
-            .Where(a => a.Length > 0)
-            .ToList();
-
-        var baseTypes = classDecl.BaseList?.Types
-            .Select(t => t.Type.ToString().Trim())
-            .Where(t => t.Length > 0)
-            .ToList() ?? new List<string>();
-
-        // The first instance constructor is our pattern source. Static constructors are skipped.
-        var ctor = classDecl.Members
-            .OfType<ConstructorDeclarationSyntax>()
-            .FirstOrDefault(c => !c.Modifiers.Any(m => m.IsKind(SyntaxKind.StaticKeyword)));
-
-        var parameters = new List<(string TypeText, string Name)>();
-        if (ctor is not null)
-        {
-            foreach (var p in ctor.ParameterList.Parameters)
-            {
-                var typeText = p.Type?.ToString().Trim() ?? string.Empty;
-                var name = p.Identifier.ValueText;
-                if (!string.IsNullOrEmpty(typeText) && !string.IsNullOrEmpty(name))
-                    parameters.Add((typeText, name));
-            }
-        }
-
-        // Collect usings so the scaffolded file can compile. We deliberately do NOT
-        // replicate the sibling's namespace declaration — the scaffolder already picks the
-        // correct namespace for the new file from the target test project.
-        var allUsings = root.Usings
-            .Select(u => u.Name?.ToString().Trim() ?? string.Empty)
-            .Where(s => s.Length > 0)
-            .ToList();
-
-        // Per scaffold-test-preview-sibling-inference-overbroad: trim usings to those
-        // actually required to resolve identifiers in the captured surface. When a Compilation
-        // is supplied AND the sibling syntax tree is part of it, we use semantic resolution to
-        // determine the required namespaces. Otherwise we fall back to keeping all usings (the
-        // pre-fix behavior) so callers without a Compilation don't silently lose usings.
-        var trimmedUsings = TrimUsingsToReferencedNamespaces(
-            classDecl, ctor, allUsings, compilation, sourceFilePath);
-
-        return new SiblingTestPattern(attributes, baseTypes, parameters, trimmedUsings, sourceFileName);
-    }
-
-    /// <summary>
-    /// Per scaffold-test-preview-sibling-inference-overbroad: returns the subset of
-    /// <paramref name="allUsings"/> that semantic resolution proves are required by identifier
-    /// references inside the captured surface — the class declaration's <c>BaseList</c> and the
-    /// constructor's parameter type list. When no Compilation is supplied (or the sibling's
-    /// syntax tree isn't part of it), returns <paramref name="allUsings"/> unchanged so
-    /// callers without semantic context don't silently drop required imports.
-    /// </summary>
-    private static IReadOnlyList<string> TrimUsingsToReferencedNamespaces(
-        ClassDeclarationSyntax classDecl,
-        ConstructorDeclarationSyntax? ctor,
-        IReadOnlyList<string> allUsings,
-        Compilation? compilation,
-        string? sourceFilePath)
-    {
-        if (compilation is null || allUsings.Count == 0)
-            return allUsings;
-
-        // Locate the syntax tree in the compilation. Match by absolute file path. If the
-        // sibling source isn't part of the compilation (e.g. a referenceTestFile pointing
-        // outside the loaded workspace), we conservatively keep all usings.
-        SyntaxTree? siblingTree = null;
-        if (!string.IsNullOrEmpty(sourceFilePath))
-        {
-            var fullPath = Path.GetFullPath(sourceFilePath);
-            siblingTree = compilation.SyntaxTrees.FirstOrDefault(t =>
-                !string.IsNullOrEmpty(t.FilePath) &&
-                string.Equals(Path.GetFullPath(t.FilePath), fullPath, StringComparison.OrdinalIgnoreCase));
-        }
-
-        if (siblingTree is null)
-            return allUsings;
-
-        var semanticModel = compilation.GetSemanticModel(siblingTree);
-
-        // Re-find the class declaration in the compilation's tree (the parsed-from-text tree
-        // we extracted above is a different syntax tree instance; semantic model needs the
-        // tree owned by the compilation).
-        var siblingRoot = siblingTree.GetCompilationUnitRoot();
-        var compilationClassDecl = siblingRoot.DescendantNodes()
-            .OfType<ClassDeclarationSyntax>()
-            .FirstOrDefault(c => string.Equals(c.Identifier.ValueText, classDecl.Identifier.ValueText, StringComparison.Ordinal)
-                              && c.SpanStart == classDecl.SpanStart);
-        if (compilationClassDecl is null)
-            return allUsings;
-
-        // Collect every identifier-name reference in the captured surface (BaseList + ctor
-        // parameter types). Resolve each to a symbol via the semantic model and pull its
-        // ContainingNamespace. The resulting set of namespace strings is what the scaffolded
-        // file needs to bring into scope to compile.
-        var referencedNamespaces = new HashSet<string>(StringComparer.Ordinal);
-
-        var surfaceNodes = new List<SyntaxNode>();
-        if (compilationClassDecl.BaseList is not null)
-            surfaceNodes.AddRange(compilationClassDecl.BaseList.Types.Select(t => t.Type));
-
-        var compilationCtor = compilationClassDecl.Members
-            .OfType<ConstructorDeclarationSyntax>()
-            .FirstOrDefault(c => !c.Modifiers.Any(m => m.IsKind(SyntaxKind.StaticKeyword)));
-        if (compilationCtor is not null)
-        {
-            foreach (var p in compilationCtor.ParameterList.Parameters)
-            {
-                if (p.Type is not null)
-                    surfaceNodes.Add(p.Type);
-            }
-        }
-
-        foreach (var node in surfaceNodes)
-        {
-            // For each type-reference node (TypeSyntax), resolve every IdentifierNameSyntax /
-            // GenericNameSyntax descendant. This catches both the outer type and any generic
-            // type arguments (e.g. IClassFixture<CustomWebApplicationFactory> resolves both
-            // IClassFixture and CustomWebApplicationFactory).
-            foreach (var nameNode in node.DescendantNodesAndSelf().OfType<SimpleNameSyntax>())
-            {
-                var typeInfo = semanticModel.GetTypeInfo(nameNode).Type
-                    ?? semanticModel.GetSymbolInfo(nameNode).Symbol as ITypeSymbol;
-                if (typeInfo is null)
-                    continue;
-                var ns = typeInfo.ContainingNamespace;
-                if (ns is null || ns.IsGlobalNamespace)
-                    continue;
-                var nsString = ns.ToDisplayString();
-                if (!string.IsNullOrEmpty(nsString))
-                    referencedNamespaces.Add(nsString);
-            }
-        }
-
-        // Keep only usings whose name matches one of the referenced namespaces. Framework
-        // usings are also filtered downstream in BuildSiblingFragments — the trim here just
-        // narrows the set to namespaces semantic resolution proved are needed.
-        return allUsings.Where(u => referencedNamespaces.Contains(u)).ToList();
-    }
-
-    private static string BuildTestContent(
-        string testNamespace,
-        ScaffoldTestDto request,
-        string simpleTypeName,
-        string targetNamespace,
-        string constructorArgs,
-        string framework,
-        IMethodSymbol? targetMethod,
-        INamedTypeSymbol? matchedType,
-        SiblingTestPattern? siblingPattern,
-        string? suggestedMethodName = null,
-        bool isTargetInaccessible = false)
-    {
-        var methodName = string.IsNullOrWhiteSpace(request.TargetMethodName)
-            ? "Generated_Test"
-            : suggestedMethodName ?? $"{request.TargetMethodName}_Needs_Test";
-
-        var usingDirective = string.IsNullOrWhiteSpace(targetNamespace)
-            ? string.Empty
-            : $"using {targetNamespace};\n";
-
-        // Collect namespaces from constructor parameter types that differ from the target's own
-        // namespace — these are not captured by the single usingDirective above and would cause
-        // CS0246 when the generated file is compiled as-is (scaffold-test-preview-missing-usings).
-        var ctorParamUsings = BuildCtorParamUsings(matchedType, targetNamespace, testNamespace);
-
-        var useStaticScaffold = ShouldUseStaticTestScaffold(matchedType);
-
-        // scaffold-test-internal-target-accessibility: when the target is internal-not-visible
-        // (or any containing type is private/internal-not-visible), skip the direct-call shape
-        // — we cannot synthesize compiling code without the caller setting up
-        // InternalsVisibleTo. Emit an Inconclusive/placeholder body and rely on the
-        // typeInfo.Warnings entry to explain the choice.
-        var ctorCall = (isTargetInaccessible || useStaticScaffold)
-            ? string.Empty
-            : string.IsNullOrWhiteSpace(constructorArgs)
-                ? $"new {simpleTypeName}()"
-                : $"new {simpleTypeName}({constructorArgs})";
-
-        var methodTargetBlock = isTargetInaccessible
-            ? BuildInaccessibleTargetPlaceholderBlock(framework, simpleTypeName, request.TargetMethodName)
-            : BuildMethodTargetInvocationBlock(
-                framework, simpleTypeName, request.TargetMethodName, targetMethod, useStaticScaffold);
-
-        // When target is inaccessible, suppress the `var subject = new T(...);` setup line
-        // — there is no callable type. The framework-specific `isStaticType=true` branch
-        // already produces a body that doesn't reference `subject`, so we route through it.
-        var suppressInstanceSubject = useStaticScaffold || isTargetInaccessible;
-
-        return framework switch
-        {
-            "xunit" => BuildXUnitTestContent(testNamespace, usingDirective, ctorParamUsings, simpleTypeName, methodName, ctorCall, methodTargetBlock, suppressInstanceSubject, siblingPattern),
-            "nunit" => BuildNUnitTestContent(testNamespace, usingDirective, ctorParamUsings, simpleTypeName, methodName, ctorCall, methodTargetBlock, suppressInstanceSubject, siblingPattern),
-            _ => BuildMSTestTestContent(testNamespace, usingDirective, ctorParamUsings, simpleTypeName, methodName, ctorCall, methodTargetBlock, suppressInstanceSubject, siblingPattern),
-        };
-    }
-
-    /// <summary>
-    /// scaffold-test-internal-target-accessibility: emits a placeholder body when the target
-    /// type/method is not visible to the test assembly. Replaces what would have been
-    /// <c>subject.M()</c> or <c>Type.M()</c> calls — those would compile-fail with CS0122.
-    /// The accompanying warning (see <see cref="BuildInaccessibleTypeWarning"/> /
-    /// <see cref="BuildInaccessibleMethodWarning"/>) explains the choice and points at
-    /// <c>InternalsVisibleTo</c>.
-    /// </summary>
-    private static string BuildInaccessibleTargetPlaceholderBlock(string framework, string targetTypeName, string? targetMethodName)
-    {
-        var inconclusiveCall = framework switch
-        {
-            "xunit" => "Assert.Fail(\"" + InaccessibleTargetAssertReason(targetTypeName, targetMethodName) + "\");",
-            "nunit" => "Assert.Inconclusive(\"" + InaccessibleTargetAssertReason(targetTypeName, targetMethodName) + "\");",
-            _ => "Assert.Inconclusive(\"" + InaccessibleTargetAssertReason(targetTypeName, targetMethodName) + "\");",
-        };
-        var methodFragment = string.IsNullOrWhiteSpace(targetMethodName)
-            ? string.Empty
-            : "." + targetMethodName;
-        return
-            $"        // Target '{targetTypeName}{methodFragment}' is not visible to this test assembly\n" +
-            "        // (declared accessibility forbids a direct call). Add InternalsVisibleTo, expose\n" +
-            "        // the target publicly, or scaffold from a project that already has access.\n" +
-            "        " + inconclusiveCall + "\n";
-    }
-
-    private static string InaccessibleTargetAssertReason(string targetTypeName, string? targetMethodName)
-    {
-        var methodFragment = string.IsNullOrWhiteSpace(targetMethodName)
-            ? string.Empty
-            : "." + targetMethodName;
-        return $"Scaffolded test for '{targetTypeName}{methodFragment}' is non-applicable: target is not accessible from this test assembly. See preview warnings for guidance.";
-    }
-
-    private static string? FormatMethodSignature(IMethodSymbol? method)
-    {
-        if (method is null)
-        {
-            return null;
-        }
-
-        var parameters = string.Join(", ", method.Parameters.Select(p => $"{p.Type.ToMinimalDisplay()} {p.Name}"));
-        return $"{method.ReturnType.ToMinimalDisplay()} {method.Name}({parameters})";
-    }
-
-    private static IReadOnlyList<string> CollectSiblingTestMethodNames(
-        string projectDirectory,
-        string destinationFilePath,
-        int maxNames)
-    {
-        if (!Directory.Exists(projectDirectory))
-        {
-            return Array.Empty<string>();
-        }
-
-        var destinationNormalized = Path.GetFullPath(destinationFilePath);
-        var names = new List<string>();
-        foreach (var file in Directory.EnumerateFiles(projectDirectory, "*Tests.cs", SearchOption.AllDirectories)
-                     .Where(p =>
-                     {
-                         var normalized = Path.GetFullPath(p);
-                         if (string.Equals(normalized, destinationNormalized, StringComparison.OrdinalIgnoreCase))
-                         {
-                             return false;
-                         }
-
-                         var rel = Path.GetRelativePath(projectDirectory, normalized);
-                         return !rel.Split(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
-                             .Any(seg => string.Equals(seg, "obj", StringComparison.OrdinalIgnoreCase)
-                                      || string.Equals(seg, "bin", StringComparison.OrdinalIgnoreCase));
-                     })
-                     .Select(p => new FileInfo(p))
-                     .OrderByDescending(fi => fi.LastWriteTimeUtc))
-        {
-            try
-            {
-                var tree = CSharpSyntaxTree.ParseText(File.ReadAllText(file.FullName));
-                var root = tree.GetRoot();
-                foreach (var method in root.DescendantNodes().OfType<MethodDeclarationSyntax>())
-                {
-                    if (!LooksLikeTestMethod(method))
-                    {
-                        continue;
-                    }
-
-                    names.Add(method.Identifier.Text);
-                    if (names.Count >= maxNames)
-                    {
-                        return names;
-                    }
-                }
-            }
-            catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-            {
-                continue;
-            }
-        }
-
-        return names;
-    }
-
-    private static bool LooksLikeTestMethod(MethodDeclarationSyntax method)
-        => method.AttributeLists
-            .SelectMany(static list => list.Attributes)
-            .Select(static attr => attr.Name.ToString())
-            .Any(static name =>
-                name.Contains("TestMethod", StringComparison.Ordinal) ||
-                name.Contains("Test", StringComparison.Ordinal) ||
-                name.Contains("Fact", StringComparison.Ordinal) ||
-                name.Contains("Theory", StringComparison.Ordinal));
-
-    private static string? NormalizeSuggestedTestMethodName(string? rawName)
-    {
-        if (string.IsNullOrWhiteSpace(rawName))
-        {
-            return null;
-        }
-
-        var candidate = rawName
-            .Split(['\r', '\n'], StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
-            .FirstOrDefault(line => !line.StartsWith("```", StringComparison.Ordinal));
-        if (string.IsNullOrWhiteSpace(candidate))
-        {
-            return null;
-        }
-
-        candidate = candidate.Trim('`', '"', '\'', ';', ' ');
-        var colon = candidate.LastIndexOf(':');
-        if (colon >= 0 && colon < candidate.Length - 1)
-        {
-            candidate = candidate[(colon + 1)..].Trim();
-        }
-        if (candidate.EndsWith("()", StringComparison.Ordinal))
-        {
-            candidate = candidate[..^2].Trim();
-        }
-        var paren = candidate.IndexOf('(');
-        if (paren > 0)
-        {
-            candidate = candidate[..paren].Trim();
-        }
-        var parts = candidate.Split(' ', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
-        if (parts.Length > 0)
-        {
-            candidate = parts[^1];
-        }
-
-        try
-        {
-            IdentifierValidation.ThrowIfInvalidIdentifier(candidate, "sampled test method name");
-            return candidate;
-        }
-        catch (InvalidOperationException)
-        {
-            return null;
-        }
-    }
-
-    /// <summary>
     /// Strip a dotted input (e.g. <c>"SampleLib.Hierarchy.Circle"</c>) to its last identifier
     /// segment so it can be used both as a lookup key against <see cref="Compilation.GetSymbolsWithName"/>
     /// (which indexes on the simple name) and as a C# identifier in scaffolded output. Callers
@@ -2373,404 +639,5 @@ public sealed class ScaffoldingService : IScaffoldingService
             return input;
         var lastDot = input.LastIndexOf('.');
         return lastDot < 0 ? input : input[(lastDot + 1)..];
-    }
-
-    /// <summary>
-    /// Renders the sibling-inferred class-level attribute lists and base-list / constructor
-    /// pieces on top of the default scaffold. Framework-level attributes (<c>[TestClass]</c>,
-    /// <c>[TestFixture]</c>) are injected by the per-framework builders; this helper only
-    /// replicates attributes that appear on the sibling class decl so cross-cutting
-    /// conventions (e.g. <c>[Trait("Category","Integration")]</c>) carry over. Returns all
-    /// three render fragments ready for string-interpolation into the per-framework emitter.
-    /// </summary>
-    private static (string ExtraUsings, string ExtraAttributes, string BaseClause, string CtorBlock) BuildSiblingFragments(
-        SiblingTestPattern? pattern,
-        string targetTypeName,
-        string testNamespace,
-        string usingDirective)
-    {
-        if (pattern is null)
-            return (string.Empty, string.Empty, string.Empty, string.Empty);
-
-        // Carry sibling usings, but skip any already present in the explicit target-type using
-        // directive, the test namespace itself, or matching the sibling's own namespace (we
-        // don't try to resolve that — we just carry the explicit usings list). The
-        // duplication guard uses a simple string comparison against the scaffold's
-        // already-emitted using so callers see a clean file.
-        var existingUsings = new HashSet<string>(StringComparer.Ordinal) { testNamespace };
-        if (!string.IsNullOrWhiteSpace(usingDirective))
-        {
-            // "using Foo.Bar;\n" → "Foo.Bar"
-            var candidate = usingDirective.Trim();
-            if (candidate.StartsWith("using ", StringComparison.Ordinal) && candidate.EndsWith(";", StringComparison.Ordinal))
-                existingUsings.Add(candidate[6..^1]);
-        }
-
-        var extraUsingsBuilder = new System.Text.StringBuilder();
-        foreach (var u in pattern.RequiredUsings.Where(u => existingUsings.Add(u)))
-        {
-            // Emit using with a newline so the header block stays consistent with the default
-            // scaffold style. Skip framework-specific usings (Xunit / MSTest / NUnit) — the
-            // per-framework emitters inject those explicitly.
-            if (IsFrameworkUsing(u))
-                continue;
-            extraUsingsBuilder.Append("using ").Append(u).Append(";\n");
-        }
-
-        var attributesBuilder = new System.Text.StringBuilder();
-        foreach (var attr in pattern.ClassAttributes.Where(a => !IsFrameworkClassAttribute(a)))
-        {
-            attributesBuilder.Append(attr).Append('\n');
-        }
-
-        var baseClause = pattern.BaseTypes.Count > 0
-            ? " : " + string.Join(", ", pattern.BaseTypes)
-            : string.Empty;
-
-        var ctorBlock = BuildCtorFromPattern(pattern, targetTypeName);
-
-        return (extraUsingsBuilder.ToString(), attributesBuilder.ToString(), baseClause, ctorBlock);
-    }
-
-    private static bool IsFrameworkUsing(string ns) =>
-        string.Equals(ns, "Xunit", StringComparison.Ordinal) ||
-        string.Equals(ns, "NUnit.Framework", StringComparison.Ordinal) ||
-        string.Equals(ns, "Microsoft.VisualStudio.TestTools.UnitTesting", StringComparison.Ordinal);
-
-    /// <summary>
-    /// Returns <c>true</c> when the captured class-level attribute should be filtered from
-    /// the scaffolded output. The blocklist covers two distinct categories:
-    /// <list type="bullet">
-    ///   <item><description>Framework class markers (<c>[TestClass]</c>, <c>[TestFixture]</c>) —
-    ///   the per-framework emitter unconditionally injects the canonical attribute, so carrying
-    ///   it from the sibling would double-emit it.</description></item>
-    ///   <item><description>Convention-tagging attributes (<c>[Trait]</c>, <c>[Category]</c>,
-    ///   <c>[TestCategory]</c>, <c>[Collection]</c>) — these classify the sibling fixture into a
-    ///   test bucket (e.g. <c>[Trait("Category","Playwright")]</c>) and over-broadly leak that
-    ///   classification onto an unrelated scaffold target. Per the
-    ///   <c>scaffold-test-preview-sibling-inference-overbroad</c> initiative, these attributes
-    ///   should NOT carry across — sibling-inference is narrowed to file-scoped namespace +
-    ///   base-class-name shape only.</description></item>
-    /// </list>
-    /// </summary>
-    private static bool IsFrameworkClassAttribute(string attribute)
-    {
-        // Strip the surrounding brackets and any argument list so bare-name matching works.
-        var inner = attribute.Trim();
-        if (inner.StartsWith('[')) inner = inner[1..];
-        if (inner.EndsWith(']')) inner = inner[..^1];
-        var parenIdx = inner.IndexOf('(');
-        if (parenIdx >= 0) inner = inner[..parenIdx];
-        inner = inner.Trim();
-        // Framework class markers — the per-framework emitter always injects the canonical
-        // attribute, so we filter the sibling's copy to avoid double-emit.
-        if (string.Equals(inner, "TestClass", StringComparison.Ordinal)
-            || string.Equals(inner, "TestFixture", StringComparison.Ordinal))
-            return true;
-        // Convention-tagging attributes — carrying these from a sibling fixture leaks the
-        // sibling's bucket classification (e.g. Playwright/Integration/UI) onto an unrelated
-        // scaffold target. Filter aggressively.
-        return string.Equals(inner, "Trait", StringComparison.Ordinal)
-            || string.Equals(inner, "Category", StringComparison.Ordinal)
-            || string.Equals(inner, "TestCategory", StringComparison.Ordinal)
-            || string.Equals(inner, "Collection", StringComparison.Ordinal);
-    }
-
-    /// <summary>
-    /// Emits a constructor block that accepts the same parameter shape as the sibling and
-    /// stores each argument in a `private readonly` field. Follows the xUnit
-    /// <c>IClassFixture&lt;T&gt;</c> / DI-injected-fixture convention seen in ASP.NET Core
-    /// integration test projects. When the sibling has no constructor, emits an empty block.
-    /// </summary>
-    private static string BuildCtorFromPattern(SiblingTestPattern pattern, string targetTypeName)
-    {
-        if (pattern.ConstructorParameters.Count == 0)
-            return string.Empty;
-
-        var fields = new System.Text.StringBuilder();
-        var assigns = new System.Text.StringBuilder();
-        var ctorParams = new List<string>();
-
-        foreach (var (typeText, name) in pattern.ConstructorParameters)
-        {
-            var fieldName = "_" + name;
-            fields.Append($"    private readonly {typeText} {fieldName};\n");
-            assigns.Append($"        {fieldName} = {name};\n");
-            ctorParams.Add($"{typeText} {name}");
-        }
-
-        return
-            fields.ToString() + "\n" +
-            $"    public {targetTypeName}GeneratedTests({string.Join(", ", ctorParams)})\n" +
-            "    {\n" +
-            assigns.ToString() +
-            "    }\n\n";
-    }
-
-    /// <summary>
-    /// BUG-N10: static classes, or instance classes whose only public API is static members (utility types
-    /// — e.g. <c>TenantConstants</c> with only <c>public const</c> fields, or <c>SnapshotContentHasher</c>
-    /// with only static methods/properties), should not scaffold <c>new T()</c> + instance assertions.
-    /// Considers methods, properties, fields (including consts), and events so types whose only surface
-    /// is static state still skip the <c>new T()</c> template.
-    /// </summary>
-    private static bool ShouldUseStaticTestScaffold(INamedTypeSymbol? matchedType)
-    {
-        if (matchedType is null)
-            return false;
-        if (matchedType.IsStatic)
-            return true;
-
-        // Only consider user-authored members — the implicitly-declared default ctor and compiler-emitted
-        // backing fields shouldn't tip the scales. Accessors of properties/events are tracked via their
-        // owning property/event (we treat the accessor methods as part of that member, not separately).
-        var candidateMembers = matchedType.GetMembers()
-            .Where(m => !m.IsImplicitlyDeclared)
-            .Where(m => m.Kind is SymbolKind.Method or SymbolKind.Property or SymbolKind.Field or SymbolKind.Event)
-            .Where(m => m is not IMethodSymbol method ||
-                        method.MethodKind is MethodKind.Ordinary or MethodKind.ExplicitInterfaceImplementation)
-            .ToList();
-
-        static bool IsInstanceVisible(ISymbol m) =>
-            !m.IsStatic &&
-            m.DeclaredAccessibility is Accessibility.Public or Accessibility.Internal or Accessibility.Protected;
-
-        static bool IsStaticVisible(ISymbol m) =>
-            m.IsStatic &&
-            m.DeclaredAccessibility is Accessibility.Public or Accessibility.Internal;
-
-        var hasVisibleInstance = candidateMembers.Any(IsInstanceVisible);
-        var hasVisibleStatic = candidateMembers.Any(IsStaticVisible);
-
-        return !hasVisibleInstance && hasVisibleStatic;
-    }
-
-    private static string BuildMethodTargetInvocationBlock(
-        string framework,
-        string targetTypeName,
-        string? targetMethodName,
-        IMethodSymbol? targetMethod,
-        bool useStaticScaffold)
-    {
-        // Phase 1: input gate — bail out on missing/unresolved target.
-        if (string.IsNullOrWhiteSpace(targetMethodName))
-        {
-            return "        // No target method specified.\n";
-        }
-
-        if (targetMethod is null)
-        {
-            return $"        // Target method '{targetMethodName}' was not resolved on {targetTypeName}.\n";
-        }
-
-        // Phase 2: static-only scaffold (utility-class shape) — emit `Type.Method()`.
-        if (useStaticScaffold && targetMethod.IsStatic)
-        {
-            return BuildStaticInvocation(targetTypeName, targetMethodName, targetMethod);
-        }
-
-        // Phase 3: private accessibility — reflection-based invocation OR an
-        // unreachable-from-static-scaffold comment.
-        if (targetMethod.DeclaredAccessibility == Accessibility.Private)
-        {
-            return BuildPrivateReflectionInvocation(framework, targetTypeName, targetMethodName, targetMethod, useStaticScaffold);
-        }
-
-        // Phase 4: ordinary public/internal instance method — emit `subject.Method()`.
-        return BuildInstanceInvocation(targetMethodName, targetMethod);
-    }
-
-    /// <summary>
-    /// Static-scaffold branch: <c>useStaticScaffold &amp;&amp; targetMethod.IsStatic</c>.
-    /// Emits a parameterless static call, capturing the return value when non-void.
-    /// </summary>
-    private static string BuildStaticInvocation(
-        string targetTypeName,
-        string targetMethodName,
-        IMethodSymbol targetMethod)
-    {
-        if (targetMethod.Parameters.Length == 0 && !targetMethod.ReturnsVoid)
-            return $"        _ = {targetTypeName}.{targetMethodName}();\n";
-        if (targetMethod.Parameters.Length == 0 && targetMethod.ReturnsVoid)
-            return $"        {targetTypeName}.{targetMethodName}();\n";
-        return $"        // Add arguments for static method '{targetMethodName}'.\n";
-    }
-
-    /// <summary>
-    /// Private-method branch: reflection invocation with framework-specific not-null
-    /// assertion. When the scaffold is static-only and the method is a private instance
-    /// member, return an explanatory comment instead — the scaffold has no <c>subject</c>
-    /// instance to invoke against.
-    /// </summary>
-    private static string BuildPrivateReflectionInvocation(
-        string framework,
-        string targetTypeName,
-        string targetMethodName,
-        IMethodSymbol targetMethod,
-        bool useStaticScaffold)
-    {
-        if (useStaticScaffold && !targetMethod.IsStatic)
-        {
-            return "        // Private instance method — not reachable from a static-only scaffold; test via public API or InternalsVisibleTo.\n";
-        }
-
-        var assertNotNull = framework switch
-        {
-            "xunit" => "Assert.NotNull(__method);",
-            "nunit" => "Assert.That(__method, Is.Not.Null);",
-            _ => "Assert.IsNotNull(__method);",
-        };
-        var flags = targetMethod.IsStatic
-            ? "System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic"
-            : "System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic";
-        var invokeTarget = targetMethod.IsStatic ? "null" : "subject";
-        return
-            "        // Private method — invoke via reflection (replace with InternalsVisibleTo or a public API test if preferred).\n" +
-            $"        var __method = typeof({targetTypeName}).GetMethod(\n" +
-            $"            \"{targetMethodName}\",\n" +
-            $"            {flags});\n" +
-            "        " + assertNotNull + "\n" +
-            $"        __method!.Invoke({invokeTarget}, null);\n";
-    }
-
-    /// <summary>
-    /// Public/internal instance-method branch. Emits <c>subject.Method()</c> for
-    /// parameterless methods (capturing the return value when non-void) or a
-    /// commented-out example for methods that take parameters.
-    /// </summary>
-    private static string BuildInstanceInvocation(string targetMethodName, IMethodSymbol targetMethod)
-    {
-        if (targetMethod.Parameters.Length == 0 && !targetMethod.ReturnsVoid)
-        {
-            return $"        _ = subject.{targetMethodName}();\n";
-        }
-
-        if (targetMethod.Parameters.Length == 0 && targetMethod.ReturnsVoid)
-        {
-            return $"        subject.{targetMethodName}();\n";
-        }
-
-        return
-            $"        // Target method '{targetMethodName}' has parameters — add arguments or use a wrapper.\n" +
-            $"        // Example: subject.{targetMethodName}(/* args */);\n";
-    }
-
-    private static string BuildMSTestTestContent(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string targetTypeName,
-        string methodName,
-        string ctorCall,
-        string methodBlock,
-        bool isStaticType,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, ctorBlock) =
-            BuildSiblingFragments(siblingPattern, targetTypeName, testNamespace, usingDirective);
-        var instanceSetup = isStaticType
-            ? string.Empty
-            : "        var subject = " + ctorCall + ";\n\n";
-        var tailAssert = isStaticType
-            ? "        Assert.IsTrue(true);\n"
-            : "        Assert.IsNotNull(subject);\n";
-        return
-            "using Microsoft.VisualStudio.TestTools.UnitTesting;\n" +
-            usingDirective +
-            ctorParamUsings +
-            extraUsings +
-            "\nnamespace " + testNamespace + ";\n\n" +
-            extraAttributes +
-            "[TestClass]\n" +
-            "public class " + targetTypeName + "GeneratedTests" + baseClause + "\n" +
-            "{\n" +
-            ctorBlock +
-            "    [TestMethod]\n" +
-            "    public void " + methodName + "()\n" +
-            "    {\n" +
-            instanceSetup +
-            methodBlock +
-            tailAssert +
-            "    }\n" +
-            "}\n";
-    }
-
-    private static string BuildXUnitTestContent(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string targetTypeName,
-        string methodName,
-        string ctorCall,
-        string methodBlock,
-        bool isStaticType,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, ctorBlock) =
-            BuildSiblingFragments(siblingPattern, targetTypeName, testNamespace, usingDirective);
-        var instanceSetup = isStaticType
-            ? string.Empty
-            : "        var subject = " + ctorCall + ";\n\n";
-        var tailAssert = isStaticType
-            ? "        Assert.True(true);\n"
-            : "        Assert.NotNull(subject);\n";
-        return
-            "using Xunit;\n" +
-            usingDirective +
-            ctorParamUsings +
-            extraUsings +
-            "\nnamespace " + testNamespace + ";\n\n" +
-            extraAttributes +
-            "public class " + targetTypeName + "GeneratedTests" + baseClause + "\n" +
-            "{\n" +
-            ctorBlock +
-            "    [Fact]\n" +
-            "    public void " + methodName + "()\n" +
-            "    {\n" +
-            instanceSetup +
-            methodBlock +
-            tailAssert +
-            "    }\n" +
-            "}\n";
-    }
-
-    private static string BuildNUnitTestContent(
-        string testNamespace,
-        string usingDirective,
-        string ctorParamUsings,
-        string targetTypeName,
-        string methodName,
-        string ctorCall,
-        string methodBlock,
-        bool isStaticType,
-        SiblingTestPattern? siblingPattern)
-    {
-        var (extraUsings, extraAttributes, baseClause, ctorBlock) =
-            BuildSiblingFragments(siblingPattern, targetTypeName, testNamespace, usingDirective);
-        var instanceSetup = isStaticType
-            ? string.Empty
-            : "        var subject = " + ctorCall + ";\n\n";
-        var tailAssert = isStaticType
-            ? "        Assert.That(true, Is.True);\n"
-            : "        Assert.That(subject, Is.Not.Null);\n";
-        return
-            "using NUnit.Framework;\n" +
-            usingDirective +
-            ctorParamUsings +
-            extraUsings +
-            "\nnamespace " + testNamespace + ";\n\n" +
-            extraAttributes +
-            "[TestFixture]\n" +
-            "public class " + targetTypeName + "GeneratedTests" + baseClause + "\n" +
-            "{\n" +
-            ctorBlock +
-            "    [Test]\n" +
-            "    public void " + methodName + "()\n" +
-            "    {\n" +
-            instanceSetup +
-            methodBlock +
-            tailAssert +
-            "    }\n" +
-            "}\n";
     }
 }
